# Apply `InitialOptimizations` more consistently in sorting