@nanosoldier 2022-12-21
[Your benchmark job](https://github.com/JuliaLang/julia/pull/47946#issuecomment-1360839937) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/0425751_vs_8cdb17b/report.md).