@KristofferC 2022-12-21
Performance now looks equal to 1.8 but allocates 3x more. Is that expected?