@LilithHafner 2022-12-21
The allocations in `sortperm!` should be zero but there is a type instability that caused about 16 bytes and 1 allocation in 1.8 and now causes 48 bytes and 3 allocations. The nanosoldier benchmark is basically this:
```julia
julia> list = rand(50_000);

julia> ix = collect(1:length(list));

julia> @benchmark sortperm!(x, $list; alg = InsertionSort, rev = true) setup=(x = copy($ix))
BenchmarkTools.Trial: 5 samples with 1 evaluation.
 Range (min … max):  1.068 s …   1.144 s  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.087 s              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.091 s ± 30.354 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▁  ▁          █                                         ▁  
  █▁▁█▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  1.07 s         Histogram: frequency by time        1.14 s <

 Memory estimate: 48 bytes, allocs estimate: 3.
```