@LilithHafner 2022-12-21
The performance diff across this PR seems reasonable under https://github.com/JuliaCI/BaseBenchmarks.jl/pull/305 as well